separate and move projects
