+++ .gitignore
@@ -1,5 +1,5 @@
-bin/*
-obj/*
-*.suo
-*.user
-/.vs
+**/bin
+**/obj
+**/*.suo
+**/*.user
+**/.vs
