+++ Chords.sln
@@ -1,7 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chords", "Chords.csproj", "{150FEF16-4570-4041-97ED-B028973A1BED}"
+# Visual Studio 15
+VisualStudioVersion = 15.0.26430.13
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chords", "Chords\Chords.csproj", "{150FEF16-4570-4041-97ED-B028973A1BED}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ChordImages", "ChordImages\ChordImages.csproj", "{1E121B60-4922-462D-8F06-60512197E673}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,6 +17,10 @@ Global
 		{150FEF16-4570-4041-97ED-B028973A1BED}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{150FEF16-4570-4041-97ED-B028973A1BED}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{150FEF16-4570-4041-97ED-B028973A1BED}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1E121B60-4922-462D-8F06-60512197E673}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1E121B60-4922-462D-8F06-60512197E673}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1E121B60-4922-462D-8F06-60512197E673}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1E121B60-4922-462D-8F06-60512197E673}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
