+++ Chords/Web.config
@@ -1,0 +1,44 @@
+<?xml version="1.0"?>
+
+<!-- 
+Chord Image Generator
+http://einaregilsson.com/chord-image-generator/
+
+Copyright (C) 2009-2012 Einar Egilsson [einar@einaregilsson.com]
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <http://www.gnu.org/licenses/>.
+-->
+<configuration>
+  <system.webServer>
+    <security>
+      <requestFiltering allowDoubleEscaping="true" />
+    </security>
+    <defaultDocument>
+      <files>
+        <clear />
+        <add value="index.html" />
+      </files>
+    </defaultDocument>
+    <handlers>
+      <add name="ChordHandler" verb="GET" path="*.png" type="EinarEgilsson.Chords.ChordHandler"
+           resourceType="Unspecified" />
+    </handlers>
+    <modules runAllManagedModulesForAllRequests="true">
+      <add name="RedirectModule" type="EinarEgilsson.Chords.RedirectModule" />
+    </modules>
+  </system.webServer>
+  <system.web>
+    <compilation debug="true" />
+  </system.web>
+</configuration>
